added boolean override for demo
setting "accounts_password_set_max_life_existing: false" for the CIS and
STIG profiles will prevent a forced password change for users with a
password which hasn't been changed in the profile's designated period of
time.  this is needed currently because the image being used to create
node1 has a "rhel" user with a password outside this period, so after
enforcing the profiles, AAP can no longer log in.