initial commit of a new EE build process
currently we rely on EE images that are maintained in external repos,
which requires coordination and time in order to make updates to our
EE images.  this is an initial take on adding a local build process,
which will allow us more granular control over any EE images we want
to use with the demo content.

note that one of the requirements for using the redhat.openshift
collection is the openshift-clients binary RPM.  adding it to the repo
works but it slows down clone operations, so it may be worth using
git-lfs for large file management.  using git-lfs would mean that
developers would need to install it locally, so we should probably
discuss it before we move to that model.