changed requirements.yml
collections are all in the Product Demos EE, so having them in the
collections/requirements.yml was redundant.  by removing the definitions
we can re-enable project collection syncing while keeping the project
sync fast in AAP.

doing so breaks the pre-commit ansible-lint hook, though, as it needs a
requirements.yml file to install any collections locally that aren't
available.  ansible-lint will look at the tests/requirements.yml file
and install any collections defined there, so that file is now a link to
the requirements file used to build the Product Demos EE.