Add availability zone mapping for VPC subnet
Occasionally the amazon.aws.ec2_vpc_subnet module would randomly choose
an availability zone where not all instance types are availble, causing
the cloud stack workflow to fail.  This PR adds a mapping of common AZs
to the regions available in the survey attached ot the Create VPC job
template, and only creates a subnet from the list of AZs.